# openapi.json
{
  "components": {
    "schemas": {
      "ApplicationCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "external_reference": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Reference"
          },
          "features": {
            "$ref": "#/components/schemas/LoanApplicationFeatures"
          }
        },
        "required": [
          "features"
        ],
        "title": "ApplicationCreateRequest",
        "type": "object"
      },
      "ApplicationListItemResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by_id": {
            "title": "Created By Id",
            "type": "string"
          },
          "current_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Decision"
          },
          "external_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Reference"
          },
          "features": {
            "$ref": "#/components/schemas/LoanApplicationFeatures"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "latest_prediction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Prediction Id"
          },
          "risk_flag": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Flag"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          }
        },
        "required": [
          "id",
          "external_reference",
          "features",
          "created_by_id",
          "created_at",
          "status"
        ],
        "title": "ApplicationListItemResponse",
        "type": "object"
      },
      "ApplicationListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApplicationListItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ApplicationListResponse",
        "type": "object"
      },
      "ApplicationResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by_id": {
            "title": "Created By Id",
            "type": "string"
          },
          "external_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Reference"
          },
          "features": {
            "$ref": "#/components/schemas/LoanApplicationFeatures"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "external_reference",
          "features",
          "created_by_id",
          "created_at"
        ],
        "title": "ApplicationResponse",
        "type": "object"
      },
      "ApplicationReviewResponse": {
        "additionalProperties": false,
        "properties": {
          "application": {
            "$ref": "#/components/schemas/ApplicationResponse"
          },
          "current_feedback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FeedbackResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "feedback_history": {
            "items": {
              "$ref": "#/components/schemas/FeedbackResponse"
            },
            "title": "Feedback History",
            "type": "array"
          },
          "latest_prediction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PredictionResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "application",
          "latest_prediction",
          "feedback_history",
          "current_feedback"
        ],
        "title": "ApplicationReviewResponse",
        "type": "object"
      },
...
      "ApplicationStatus": {
        "enum": [
          "unscored",
          "scored",
          "decided",
          "escalated"
        ],
        "title": "ApplicationStatus",
        "type": "string"
      },
...
      "ExplanationResponse": {
        "additionalProperties": false,
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "narrative": {
            "title": "Narrative",
            "type": "string"
          },
          "top_features": {
            "items": {
              "$ref": "#/components/schemas/FeatureContribution"
            },
            "title": "Top Features",
            "type": "array"
          }
        },
        "required": [
          "available",
          "narrative"
        ],
        "title": "ExplanationResponse",
        "type": "object"
      },
      "ExplanationVariant": {
        "enum": [
          "score_only",
          "explanation_shown"
        ],
        "title": "ExplanationVariant",
        "type": "string"
      },
...
      "FeedbackResponse": {
        "additionalProperties": false,
        "properties": {
          "agreed_with_model": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agreed With Model"
          },
          "amends_feedback_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amends Feedback Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "$ref": "#/components/schemas/OfficerDecision"
          },
          "detector_state": {
            "$ref": "#/components/schemas/DriftStatusResponse"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_current": {
            "title": "Is Current",
            "type": "boolean"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "officer_id": {
            "title": "Officer Id",
            "type": "string"
          },
          "prediction_id": {
            "title": "Prediction Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "prediction_id",
          "officer_id",
          "version",
          "amends_feedback_id",
          "is_current",
          "decision",
          "agreed_with_model",
          "note",
          "detector_state",
          "created_at"
        ],
        "title": "FeedbackResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LoanApplicationFeatures": {
        "additionalProperties": false,
        "description": "Strict serving schema aligned with the frozen LendingClub transformer.",
        "properties": {
          "annual_inc": {
            "anyOf": [
              {
                "maximum": 1000000000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annual Inc"
          },
          "delinq_2yrs": {
            "anyOf": [
              {
                "maximum": 10000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delinq 2Yrs"
          },
          "dti": {
            "anyOf": [
              {
                "maximum": 1000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dti"
          },
          "earliest_cr_line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Earliest Cr Line"
          },
          "emp_length": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emp Length"
          },
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "home_ownership": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Ownership"
          },
          "inq_last_6mths": {
            "anyOf": [
              {
                "maximum": 10000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inq Last 6Mths"
          },
          "installment": {
            "anyOf": [
              {
                "maximum": 10000000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Installment"
          },
          "int_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Int Rate"
          },
          "issue_d": {
            "title": "Issue D",
            "type": "string"
          },
          "loan_amnt": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 10000000.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loan Amnt"
          },
          "open_acc": {
            "anyOf": [
              {
                "maximum": 10000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open Acc"
          },
          "purpose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purpose"
          },
          "revol_bal": {
            "anyOf": [
              {
                "maximum": 1000000000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revol Bal"
          },
          "revol_util": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revol Util"
          },
          "sub_grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Grade"
          },
          "term": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Term"
          },
          "total_acc": {
            "anyOf": [
              {
                "maximum": 10000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Acc"
          }
        },
        "required": [
          "annual_inc",
          "dti",
          "revol_util",
          "revol_bal",
          "open_acc",
          "total_acc",
          "delinq_2yrs",
          "inq_last_6mths",
          "loan_amnt",
          "term",
          "int_rate",
          "installment",
          "grade",
          "sub_grade",
          "purpose",
          "emp_length",
          "home_ownership",
          "earliest_cr_line",
          "issue_d"
        ],
        "title": "LoanApplicationFeatures",
        "type": "object"
      },
...
      "OfficerDecision": {
        "enum": [
          "approve",
          "decline",
          "escalate"
        ],
        "title": "OfficerDecision",
        "type": "string"
      },
...
      "PredictionResponse": {
        "additionalProperties": false,
        "properties": {
          "application_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "detector_state": {
            "$ref": "#/components/schemas/DriftStatusResponse"
          },
          "explanation": {
            "$ref": "#/components/schemas/ExplanationResponse"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "requested_by_id": {
            "title": "Requested By Id",
            "type": "string"
          },
          "risk_flag": {
            "title": "Risk Flag",
            "type": "boolean"
          },
          "score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score",
            "type": "number"
          },
          "study_variant": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExplanationVariant"
              },
              {
                "type": "null"
              }
            ]
          },
          "threshold": {
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Threshold",
            "type": "number"
          }
        },
        "required": [
          "id",
          "application_id",
          "requested_by_id",
          "model_version",
          "score",
          "threshold",
          "risk_flag",
          "explanation",
          "detector_state",
          "created_at"
        ],
        "title": "PredictionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Human-in-the-loop loan default decision support with auditable model scores, explanations, officer feedback, and drift state. Decision support only; a qualified human makes the final decision.",
    "title": "Drift Aware Loan Default API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/applications": {
      "get": {
        "description": "Return a role-aware, server-paginated review queue.",
        "operationId": "list_applications_api_v1_applications_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ApplicationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "risk_flag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Flag"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Applications",
        "tags": [
          "applications"
        ]
      },
      "post": {
        "operationId": "create_application_api_v1_applications_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Application",
        "tags": [
          "applications"
        ]
      }
    },
    "/api/v1/applications/{application_id}": {
      "get": {
        "operationId": "get_application_api_v1_applications__application_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Application",
        "tags": [
          "applications"
        ]
      }
    },
    "/api/v1/applications/{application_id}/review": {
      "get": {
        "description": "Fetch the application, latest score, and append-only decision history.",
        "operationId": "get_application_review_api_v1_applications__application_id__review_get",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Application Review",
        "tags": [
          "applications"
        ]
      }
    },
...
  }
}